backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
minio==7.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
Pillow==10.1.0
aiofiles==23.2.1 
email_validator==2.2.0
openpyxl==3.1.5
xlrd==2.0.1
reportlab==4.0.7
matplotlib==3.8.2
APScheduler==3.10.4